+++ cores/arduino/Stream.h
@@ -59,7 +59,7 @@ class Stream : public Print
   public:
     virtual int available() = 0;
     virtual int read() = 0;
-    virtual size_t read(uint8_t *buffer, size_t size);
+    virtual int read(uint8_t *buffer, size_t size);
     virtual int peek() = 0;
 
     Stream() {_timeout=1000;}
